server/requirements.txt
django==4.2.25
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-allauth==65.12.1
dj-rest-auth==7.0.1
djangorestframework-simplejwt==5.5.1
pymysql==1.1.1
python-dotenv==1.1.1
requests==2.32.5
cryptography==46.0.3
